Moved translation functionality to separate file, i18n.js
