+++ smart_card_connector_app/src/window-main.js
@@ -24,6 +24,7 @@ goog.require('GoogleSmartCard.ConnectorApp.Window.AboutShowing');
 goog.require('GoogleSmartCard.ConnectorApp.Window.AppsDisplaying');
 goog.require('GoogleSmartCard.ConnectorApp.Window.DevicesDisplaying');
 goog.require('GoogleSmartCard.ConnectorApp.Window.LogsExporting');
+goog.require('GoogleSmartCard.I18n');
 goog.require('GoogleSmartCard.Logging');
 goog.require('GoogleSmartCard.PopupWindow.Client');
 goog.require('goog.dom');
@@ -60,52 +61,7 @@ GSC.ConnectorApp.Window.LogsExporting.initialize();
 
 GSC.PopupWindow.Client.showWindow();
 
-/** @const */
-var I18N_DATA_ATTRIBUTE = 'data-i18n';
-
-/** @const */
-var I18N_DATA_ARIA_LABEL_ATTRIBUTE = 'data-i18n-aria-label';
-
-function setElementI18nText(element) {
-  var i18nId = element.getAttribute(I18N_DATA_ATTRIBUTE);
-  if (!i18nId) {
-    GSC.Logging.failWithLogger(
-        logger,
-        'Failed to get element attribute: ' + element.outerHTML);
-  }
-  var translatedText = chrome.i18n.getMessage(i18nId);
-  if (!translatedText) {
-    GSC.Logging.failWithLogger(
-        logger,
-        'Failed to get translation for text with id: ' + i18nId);
-  }
-  element.textContent = translatedText;
-}
-
-function setElementI18nAriaLabel(element) {
-  var i18nId = element.getAttribute(I18N_DATA_ARIA_LABEL_ATTRIBUTE);
-  if (!i18nId) {
-    GSC.Logging.failWithLogger(
-        logger,
-        'Failed to get element attribute: ' + element.outerHTML);
-  }
-  var translatedText = chrome.i18n.getMessage(i18nId);
-  if (!translatedText) {
-    GSC.Logging.failWithLogger(
-        logger,
-        'Failed to get translation for text with id: ' + i18nId);
-  }
-  element.setAttribute('aria-label', translatedText);
-}
-
-for (let element of document.querySelectorAll(
-    '[' + I18N_DATA_ATTRIBUTE + ']')) {
-  setElementI18nText(element);
-}
-
-for (let element of document.querySelectorAll(
-    '[' + I18N_DATA_ARIA_LABEL_ATTRIBUTE + ']')) {
-  setElementI18nAriaLabel(element);
-}
+// This call does all of the translation work.
+GSC.I18n.adjustElementsTranslation();
 
 });  // goog.scope
